requirements.txt
numpy
pandas
matplotlib
scikit-learn
seaborn 
requests
flask
mysql-connector-python 
pymysql
python-dotenv


-e . #For Local Package Installation 
